Restore low-cost working spinner
