+++ src/renderer/src/components/sidebar/StatusIndicator.test.ts
@@ -17,11 +17,12 @@ function renderDotClassNames(status: Status): string[] {
 }
 
 describe('StatusIndicator', () => {
-  it('renders working as a static yellow ring', () => {
+  it('renders working as a stepped yellow spinner', () => {
     const classNames = renderDotClassNames('working')
 
     expect(classNames).toContain('border-yellow-500')
-    expect(classNames).toContain('bg-yellow-500/15')
+    expect(classNames).toContain('border-t-transparent')
+    expect(classNames).toContain('[animation:spin_1s_steps(12,end)_infinite]')
     expect(classNames).not.toContain('animate-spin')
   })
